+++ ThirdPartyAdapters/mopub/mopub/src/main/java/com/mopub/mobileads/dfp/adapters/MoPubAdapter.java
@@ -4,11 +4,13 @@ import android.app.Activity;
 import android.content.Context;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
+import android.text.TextUtils;
 import android.text.format.DateFormat;
 import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 
+import com.google.ads.mediation.mopub.MoPubSingleton;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdSize;
 import com.google.android.gms.ads.formats.NativeAdOptions;
@@ -38,6 +40,7 @@ import com.mopub.nativeads.ViewBinder;
 
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.EnumSet;
@@ -59,7 +62,8 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
     private AdSize mAdSize;
 
     private MoPubInterstitial mMoPubInterstitial;
-    private static final String MOPUB_NATIVE_CEVENT_VERSION = "gmext";
+    private MediationInterstitialListener mMediationInterstitialListener;
+    public static final String MOPUB_NATIVE_CEVENT_VERSION = "gmext";
     public static final double DEFAULT_MOPUB_IMAGE_SCALE = 1;
     private static final String MOPUB_AD_UNIT_KEY = "adUnitId";
     private int privacyIconPlacement;
@@ -74,12 +78,10 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
 
     @Override
     public void onDestroy() {
-
         if (mMoPubInterstitial != null) {
             mMoPubInterstitial.destroy();
             mMoPubInterstitial = null;
         }
-
         if (mMoPubView != null) {
             mMoPubView.destroy();
             mMoPubView = null;
@@ -101,7 +103,12 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
                                 NativeMediationAdRequest mediationAdRequest,
                                 Bundle mediationExtras) {
 
-        String adunit = serverParameters.getString(MOPUB_AD_UNIT_KEY);
+        String adUnit = serverParameters.getString(MOPUB_AD_UNIT_KEY);
+        if (TextUtils.isEmpty(adUnit)) {
+            Log.d(TAG, "Missing or Invalid MoPub Ad Unit ID.");
+            listener.onAdFailedToLoad(MoPubAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
+            return;
+        }
 
         final NativeAdOptions options = mediationAdRequest.getNativeAdOptions();
 
@@ -223,13 +230,7 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
                     }
                 };
 
-        if (adunit == null) {
-            Log.d(TAG, "Ad unit id is invalid. So failing the request.");
-            listener.onAdFailedToLoad(MoPubAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
-            return;
-        }
-
-        MoPubNative moPubNative = new MoPubNative(context, adunit, moPubNativeNetworkListener);
+        final MoPubNative moPubNative = new MoPubNative(context, adUnit, moPubNativeNetworkListener);
 
         ViewBinder viewbinder = new ViewBinder.Builder(0).build();
         MoPubStaticNativeAdRenderer moPubStaticNativeAdRenderer =
@@ -249,11 +250,14 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
                 .desiredAssets(assetsSet)
                 .build();
 
-        if (MoPub.isSdkInitialized()) {
-            moPubNative.makeRequest(requestParameters);
-        } else {
-            initializeMoPub(context, adunit, null, null, moPubNative);
-        }
+        SdkConfiguration sdkConfiguration = new SdkConfiguration.Builder(adUnit).build();
+        MoPubSingleton.getInstance().initializeMoPubSDK(context, sdkConfiguration,
+                new SdkInitializationListener() {
+            @Override
+            public void onInitializationFinished() {
+                moPubNative.makeRequest(requestParameters);
+            }
+        });
 
         // Forwarding MoPub's impression and click events to AdMob.
         mMoPubNativeEventListener = new NativeAd.MoPubNativeEventListener() {
@@ -282,12 +286,24 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
                                 MediationAdRequest mediationAdRequest,
                                 Bundle bundle1) {
 
-        String adunit = bundle.getString(MOPUB_AD_UNIT_KEY);
+        String adUnit = bundle.getString(MOPUB_AD_UNIT_KEY);
+        if (TextUtils.isEmpty(adUnit)) {
+            Log.d(TAG, "Missing or Invalid MoPub Ad Unit ID.");
+            mediationBannerListener.onAdFailedToLoad(MoPubAdapter.this,
+                    AdRequest.ERROR_CODE_INVALID_REQUEST);
+            return;
+        }
 
-        mAdSize = adSize;
+        mAdSize = getSupportedAdSize(context, adSize);
+        if (mAdSize == null) {
+            Log.w(TAG, "Failed to request ad, AdSize is null.");
+            mediationBannerListener.onAdFailedToLoad(this, AdRequest.ERROR_CODE_INVALID_REQUEST);
+            return;
+        }
+
         mMoPubView = new MoPubView(context);
         mMoPubView.setBannerAdListener(new MBannerListener(mediationBannerListener));
-        mMoPubView.setAdUnitId(adunit);
+        mMoPubView.setAdUnitId(adUnit);
 
         //If test mode is enabled
         if (mediationAdRequest.isTesting()) {
@@ -302,12 +318,86 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
         mMoPubView.setKeywords(getKeywords(mediationAdRequest, false));
         mMoPubView.setUserDataKeywords(getKeywords(mediationAdRequest, true));
 
-        if (MoPub.isSdkInitialized()) {
-            mMoPubView.loadAd();
-        } else {
-            initializeMoPub(context, adunit, mMoPubView, null, null);
-        }
+        SdkConfiguration sdkConfiguration = new SdkConfiguration.Builder(adUnit).build();
+        MoPubSingleton.getInstance().initializeMoPubSDK(context, sdkConfiguration,
+                new SdkInitializationListener() {
+            @Override
+            public void onInitializationFinished() {
+               mMoPubView.loadAd();
+            }
+        });
     }
+
+    private AdSize getSupportedAdSize(Context context, AdSize adSize) {
+        AdSize original = new AdSize(adSize.getWidth(),
+                adSize.getHeight());
+
+        ArrayList<AdSize> potentials = new ArrayList<>(2);
+        potentials.add(AdSize.BANNER);
+        potentials.add(AdSize.MEDIUM_RECTANGLE);
+        potentials.add(AdSize.LEADERBOARD);
+        potentials.add(AdSize.WIDE_SKYSCRAPER);
+        Log.i(TAG, potentials.toString());
+        return findClosestSize(context, original, potentials);
+    }
+
+    // Start of helper code to remove when available in SDK
+    /**
+     * Find the closest supported AdSize from the list of potentials to the provided size.
+     * Returns null if none are within given threshold size range.
+     */
+    public static AdSize findClosestSize(
+            Context context, AdSize original, ArrayList<AdSize> potentials) {
+        if (potentials == null || original == null) {
+            return null;
+        }
+        float density = context.getResources().getDisplayMetrics().density;
+        int actualWidth = Math.round(original.getWidthInPixels(context)/density);
+        int actualHeight = Math.round(original.getHeightInPixels(context)/density);
+        original = new AdSize(actualWidth, actualHeight);
+        AdSize largestPotential = null;
+        for (AdSize potential : potentials) {
+            if (isSizeInRange(original, potential)) {
+                if (largestPotential == null) {
+                    largestPotential = potential;
+                } else {
+                    largestPotential = getLargerByArea(largestPotential, potential);
+                }
+            }
+        }
+        return largestPotential;
+    }
+
+    private static boolean isSizeInRange(AdSize original, AdSize potential) {
+        if (potential == null) {
+            return false;
+        }
+        double minWidthRatio = 0.5;
+        double minHeightRatio = 0.7;
+
+        int originalWidth = original.getWidth();
+        int potentialWidth = potential.getWidth();
+        int originalHeight = original.getHeight();
+        int potentialHeight = potential.getHeight();
+
+        if (originalWidth * minWidthRatio > potentialWidth ||
+                originalWidth < potentialWidth) {
+            return false;
+        }
+
+        if (originalHeight * minHeightRatio > potentialHeight ||
+                originalHeight < potentialHeight) {
+            return false;
+        }
+        return true;
+    }
+
+    private static AdSize getLargerByArea(AdSize size1, AdSize size2) {
+        int area1 = size1.getWidth() * size1.getHeight();
+        int area2 = size2.getWidth() * size2.getHeight();
+        return area1 > area2 ? size1 : size2;
+    }
+    // End code to remove when available in SDK
 
     @Override
     public View getBannerView() {
@@ -421,8 +511,10 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
         public void onBannerLoaded(MoPubView moPubView) {
             if (!(mAdSize.getWidth() == moPubView.getAdWidth()
                     && mAdSize.getHeight() == moPubView.getAdHeight())) {
-                Log.w(TAG, "The banner ad size loaded does not match the request size. Update the"
+                Log.e(TAG, "The banner ad size loaded does not match the request size. Update the"
                         + " ad size on your MoPub UI to match the request size.");
+                mMediationBannerListener.onAdFailedToLoad(MoPubAdapter.this, AdRequest.ERROR_CODE_NO_FILL);
+                return;
             }
             mMediationBannerListener.onAdLoaded(MoPubAdapter.this);
 
@@ -436,11 +528,26 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
                                       MediationAdRequest mediationAdRequest,
                                       Bundle bundle1) {
 
-        String adunit = bundle.getString(MOPUB_AD_UNIT_KEY);
+        if (!(context instanceof Activity)) {
+            Log.w(TAG, "MoPub SDK requires an Activity context to load interstitial ads.");
+            mediationInterstitialListener.onAdFailedToLoad(MoPubAdapter.this,
+                    AdRequest.ERROR_CODE_INVALID_REQUEST);
+            return;
+        }
 
-        mMoPubInterstitial = new MoPubInterstitial((Activity) context, adunit);
+        String adUnit = bundle.getString(MOPUB_AD_UNIT_KEY);
+        if (TextUtils.isEmpty(adUnit)) {
+            Log.d(TAG, "Missing or Invalid MoPub Ad Unit ID.");
+            mediationInterstitialListener.onAdFailedToLoad(MoPubAdapter.this,
+                    AdRequest.ERROR_CODE_INVALID_REQUEST);
+            return;
+        }
+
+        mMediationInterstitialListener = mediationInterstitialListener;
+
+        mMoPubInterstitial = new MoPubInterstitial((Activity) context, adUnit);
         mMoPubInterstitial.setInterstitialAdListener(
-                new mMediationInterstitialListener(mediationInterstitialListener));
+                new mMediationInterstitialListener(mMediationInterstitialListener));
 
         //If test mode is enabled
         if (mediationAdRequest.isTesting()) {
@@ -450,20 +557,26 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
         mMoPubInterstitial.setKeywords(getKeywords(mediationAdRequest, false));
         mMoPubInterstitial.setKeywords(getKeywords(mediationAdRequest, true));
 
-        if (MoPub.isSdkInitialized()) {
-            mMoPubInterstitial.load();
-        } else {
-            initializeMoPub(context, adunit, null, mMoPubInterstitial, null);
-        }
+        SdkConfiguration sdkConfiguration = new SdkConfiguration.Builder(adUnit).build();
+        MoPubSingleton.getInstance().initializeMoPubSDK(context, sdkConfiguration,
+                new SdkInitializationListener() {
+            @Override
+            public void onInitializationFinished() {
+                mMoPubInterstitial.load();
+            }
+        });
     }
 
     @Override
     public void showInterstitial() {
-
         if (mMoPubInterstitial.isReady()) {
             mMoPubInterstitial.show();
         } else {
             MoPubLog.i("Interstitial was not ready. Unable to load the interstitial");
+            if (mMediationInterstitialListener != null) {
+                mMediationInterstitialListener.onAdOpened(MoPubAdapter.this);
+                mMediationInterstitialListener.onAdClosed(MoPubAdapter.this);
+            }
         }
     }
 
@@ -503,6 +616,12 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
                         mMediationInterstitialListener.onAdFailedToLoad(MoPubAdapter.this,
                                 AdRequest.ERROR_CODE_INVALID_REQUEST);
                         break;
+                    case EXPIRED:
+                        // MoPub Rewarded video ads expire after 4 hours.
+                        Log.i(TAG, "The MoPub Ad has expired. Please make a new Ad Request.");
+                        mMediationInterstitialListener.onAdFailedToLoad(MoPubAdapter.this,
+                                AdRequest.ERROR_CODE_NO_FILL);
+                        break;
                     default:
                         mMediationInterstitialListener.onAdFailedToLoad(MoPubAdapter.this,
                                 AdRequest.ERROR_CODE_INTERNAL_ERROR);
@@ -521,38 +640,6 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
         public void onInterstitialShown(MoPubInterstitial moPubInterstitial) {
             mMediationInterstitialListener.onAdOpened(MoPubAdapter.this);
         }
-    }
-
-    // Initializing the MoPub SDK. Required as of 5.0.0
-    private void initializeMoPub(Context context, String adUnitId, MoPubView bannerAd, MoPubInterstitial
-            interstitialAd, MoPubNative nativeAd) {
-        SdkConfiguration sdkConfiguration = new SdkConfiguration.Builder(adUnitId)
-                .build();
-        MoPub.initializeSdk(context, sdkConfiguration, initSdkListener(bannerAd, interstitialAd, nativeAd));
-    }
-
-    private SdkInitializationListener initSdkListener(final MoPubView bannerAd, final MoPubInterstitial
-            interstitialAd, final MoPubNative nativeAd) {
-        return new SdkInitializationListener() {
-
-            @Override
-            public void onInitializationFinished() {
-                MoPubLog.d("MoPub SDK initialized.");
-
-                // Sending ad requests now that the MoPub SDK has initialized
-                if (bannerAd != null) {
-                    bannerAd.loadAd();
-                } else if (interstitialAd != null) {
-                    interstitialAd.load();
-                } else if (nativeAd != null) {
-                    if (requestParameters != null) {
-                        nativeAd.makeRequest(requestParameters);
-                    } else {
-                        nativeAd.makeRequest();
-                    }
-                }
-            }
-        };
     }
 
     /**