+++ ThirdPartyAdapters/mytarget/CHANGELOG.md
@@ -1,5 +1,20 @@
 # myTarget Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 5.4.5.0
+- Verified compatibility with myTarget SDK version 5.4.5.
+- Updated the minimum required Google Mobile Ads SDK version to 18.1.0.
+
+## 5.4.0.0
+- Verified compatibility with myTarget SDK version 5.4.0.
+
+## 5.3.9.1
+- Added support for flexible banner ad sizes.
+
+## 5.3.9.0
+- Verified compatibility with myTarget SDK version 5.3.9.
+- Updated adapter to support new open-beta Rewarded API.
+- Updated the minimum required Google Mobile Ads SDK version to 17.2.0.
+
 ## 5.3.6.0
 - Verified compatibility with myTarget SDK version 5.3.6.